requirements.txt
instaloader
pandas
Pillow
google-genai
python-dotenv
tqdm
requests
